\name{demopat}
\alias{demopat}
\docType{data}
\title{
   Artificial Data Point Pattern
}
\description{
  This is an artificial dataset, for use in testing and demonstrating the
  capabilities of the \code{spatstat} package.
  It is a multitype point pattern in an irregular polygonal window.
  There are two types of points. The window contains a polygonal hole.
} 
\format{
  An object of class \code{"ppp"}
  representing the point pattern.

  See \code{\link{ppp.object}} for details of the format of a
  point pattern object.
}
\usage{data(demopat)}
\source{Adrian Baddeley \email{Adrian.Baddeley@curtin.edu.au}}
\keyword{datasets}
\keyword{spatial}


 
