\name{dim.dppmodel}
\alias{dim.dppmodel}
\title{Dimension of Determinantal Point Process Model}
\description{Extracts the dimension of a determinantal point process model.}
\usage{
  \method{dim}{dppmodel}(x)
}
\arguments{
  \item{x}{object of class \code{"dppmodel"}.}
}
\value{A numeric (or NULL if the dimension of the model is unspecified).}
\author{
  Adrian Baddeley
  \email{Adrian.Baddeley@uwa.edu.au}
  \url{http://www.maths.uwa.edu.au/~adrian/}
  Rolf Turner
  \email{r.turner@auckland.ac.nz}
  and Ege Rubak
  \email{rubak@math.aau.dk}
}
\keyword{spatial}
\keyword{models}
