\name{foo}
\alias{foo}
\alias{plot.foo}
\title{
  Foo is Not a Real Name
}
\description{
  The name \code{foo} is not a real name: it is a place holder,
  used to represent the name of any desired thing.

  The functions defined here simply print an explanation of the
  placeholder name \code{foo}.
}
\usage{
foo()

\method{plot}{foo}(x, \dots)
}
\arguments{
  \item{x}{Ignored.}
  \item{\dots}{Ignored.}
}
\details{
  The name \code{foo} is used by computer scientists as a
  \emph{place holder}, to represent the name of any desired object or
  function. It is not the name of an actual object or function;
  it serves only as an example, to explain a concept.

  However, many users misinterpret this convention, and actually
  type the command \code{foo} or \code{foo()}. Then they email the
  package author to inform them that \code{foo} is not defined.

  To avoid this correspondence, we have now defined an object
  called \code{foo}.
  
  The function \code{foo()} prints a message explaining that \code{foo}
  is not really the name of a variable. 

  The function can be executed simply by typing \code{foo}
  without parentheses.
}
\value{
  Null.
}
\author{
  \spatstatAuthors.
}
\seealso{
  \code{\link{beginner}}
}
\examples{
  foo
}
\keyword{documentation}
