\name{hierpair.family}
\alias{hierpair.family}
\title{Hierarchical Pairwise Interaction Process Family}
\description{
  An object describing the family of all hierarchical pairwise interaction Gibbs
  point processes.
}
\details{
  \bold{Advanced Use Only!}
  
  This structure would not normally be touched by
  the user. It describes 
  the hierarchical pairwise interaction family of point process models.
}
\value{
  Object of class \code{"isf"}, see \code{\link{isf.object}}.
}
\seealso{
  Other families: 
  \code{\link{pairwise.family}},
  \code{\link{pairsat.family}},
  \code{\link{ord.family}},
  \code{\link{inforder.family}}.

  Hierarchical Strauss interaction:
  \code{\link{HierStrauss}}.
}
\author{
  \spatstatAuthors.
}
\keyword{spatial}
\keyword{models}
