\name{hybrid.family}
\alias{hybrid.family}
\title{
  Hybrid Interaction Family
}
\description{
  An object describing the family of all hybrid interactions.
}
\details{
  \bold{Advanced Use Only!}
  
  This structure would not normally be touched by
  the user. It describes 
  the family of all hybrid point process models.
 
  If you need to create a specific hybrid interaction model for use in 
  modelling, use the function \code{\link{Hybrid}}.
}
\value{
  Object of class \code{"isf"}, see \code{\link{isf.object}}.
}
\seealso{
  Use \code{\link{Hybrid}} to make hybrid interactions.
  
  Other families: 
  \code{\link{pairwise.family}},
  \code{\link{pairsat.family}},
  \code{\link{ord.family}},
  \code{\link{inforder.family}}.
}
\author{
  \adrian
  and \rolf.
}
\keyword{spatial}
\keyword{models}
