\name{is.dppm}
\alias{is.dppm}
\title{Recognise Fitted Determinantal Point Process Models}
\description{Check that an object inherits the class dppm}
\usage{is.dppm(x)}
\arguments{
  \item{x}{Any object.}
}

\value{A single logical value.}

\author{Ege Rubak \email{rubak@math.aau.dk} <rubak@math.aau.dk>,
Adrian Baddeley \email{Adrian.Baddeley@curtin.edu.au} <Adrian.Baddeley@uwa.edu.au>
and Rolf Turner \email{r.turner@auckland.ac.nz} <r.turner@auckland.ac.nz>}

\keyword{spatial}
\keyword{manip}
\keyword{models}
