\name{is.lpp}
\alias{is.lpp}
\title{Test Whether An Object Is A Point Pattern on a Linear Network}
\description{
  Checks whether its argument is a point pattern
  on a linear network
  (object of class \code{"lpp"}).
}
\usage{
is.lpp(x)
}
\arguments{
  \item{x}{Any object.}
}
\details{
  This function tests whether the object \code{x}
  is a point pattern object of class \code{"lpp"}. 
}
\value{
  \code{TRUE} if \code{x} is a point pattern of class \code{"lpp"},
  otherwise \code{FALSE}.
}
\author{Adrian Baddeley \email{Adrian.Baddeley@curtin.edu.au}
  
  
  and Rolf Turner \email{r.turner@auckland.ac.nz}
  
}
\keyword{spatial}
\keyword{manip}
