\name{japanesepines}
\alias{japanesepines}
\docType{data}
\title{
  Japanese Pines Point Pattern
}
\description{
  The data give the locations of Japanese black pine saplings
  in a square sampling region in a natural forest.
  The observations were originally collected by Numata (1961).

  These data are used as a standard example in the textbook of 
  Diggle (2003); see pages 1, 14, 19, 22, 24, 56--57 and 61.
} 
\format{
  An object of class \code{"ppp"}
  representing the point pattern of tree locations
  in a 5.7 x 5.7 metre square, rescaled to the unit square
  and rounded to two decimal places.

  See \code{\link{ppp.object}} for details of the format of a
  point pattern object.
}
\usage{data(japanesepines)}
\source{Diggle (2003), obtained from Numata (1961)}

\references{
  Diggle, P.J. (2003)
  \emph{Statistical Analysis of Spatial Point Patterns}.
  Arnold Publishers.

  Numata, M. (1961)
  Forest vegetation in the vicinity of Choshi. Coastal flora and
  vegetation at Choshi, Chiba Prefecture. IV.
  \emph{Bulletin of Choshi Marine Laboratory, Chiba University}
  \bold{3}, 28--48 (in Japanese).
}
\keyword{datasets}
\keyword{spatial}
