\name{letterR}
\alias{letterR}
\title{Window in Shape of Letter R}
\description{
  A window in the shape of the capital letter R,
  for use in demonstrations.
} 
\format{
  An object of class \code{"owin"}
  representing the capital letter R,
  in the same font as the R package logo.
  See \code{\link{owin.object}} for details of the format.
}
\usage{
 data(letterR)
}
\source{Adrian Baddeley}
\keyword{datasets}
\keyword{spatial}
