\name{simba}
\alias{simba}
\docType{data}
\title{
  Simulated data from a two-group experiment with replication
  within each group.
}
\description{
  The \code{simba} dataset contains simulated data from an
  experiment with a `control' group and a `treatment' group, each 
  group containing 5 experimental units.

  The responses in the experiment are point patterns.
  
  The responses in the control group are 
  independent realisations of a Poisson point process with intensity 80.

  The responses in the treatment group are independent realisations of
  a Strauss process with activity parameter \eqn{\beta=100}{beta=100},
  interaction parameter \eqn{\gamma=0.5}{gamma=0.5} and
  interaction radius \eqn{R=0.07} in the unit square.
} 
\format{
  \code{simba} is a hyperframe with 10 rows, and columns named:

  \itemize{
    \item \code{Points} containing the point patterns
    \item \code{group} factor identifying the experimental group,
    with levels \code{control} and \code{treatment}).
  }
}
\usage{data(simba)}
\source{
  Simulated data, generated by Adrian Baddeley.
}
\keyword{datasets}
\keyword{spatial}
