\name{simdat}
\alias{simdat}
\docType{data}
\title{
  Simulated Point Pattern
}
\description{
  This point pattern data set was simulated (using the
  Metropolis-Hastings algorithm)
  from a model fitted to the Numata
  Japanese black pine data set referred to in
  Baddeley and Turner (2000).
} 
\format{
  An object of class \code{"ppp"}
  in a square window of size 10 by 10 units.

  See \code{\link{ppp.object}} for details of the format of a
  point pattern object.
}
\usage{data(simdat)}
\source{Rolf Turner \email{r.turner@auckland.ac.nz}}
\references{
  Baddeley, A. and Turner, R. (2000)
  Practical maximum pseudolikelihood for spatial point patterns.
  \emph{Australian and New Zealand Journal of Statistics}
  \bold{42}, 283--322.
}
\keyword{datasets}
\keyword{spatial}
