\name{spatstat-deprecated}
\alias{conspire}
\alias{mpl}
\title{Deprecated spatstat functions}
\description{
  Deprecated spatstat functions.
}
\usage{
conspire(\dots)
mpl(Q, trend, interaction, data, correction, rbord, use.gam)
}

\details{
  \code{mpl} has been replaced by \code{\link{ppm}}.

  \code{conspire} has been replaced by \code{\link{plot.fv}}.
  
}
\keyword{internal}
