\name{spatstat-deprecated}
\alias{conspire}
\alias{ksmooth.ppp}
\alias{mpl}
\title{Deprecated spatstat functions}
\description{
  Deprecated spatstat functions.
}
\usage{
conspire(\dots)
ksmooth.ppp(x, sigma, \dots, edge=TRUE)
mpl(Q, trend, interaction, data, correction, rbord, use.gam)
}

\details{
  \code{conspire} has been replaced by \code{\link{plot.fv}}.

  \code{ksmooth.ppp} has been replaced by \code{\link{density.ppp}}.

  \code{mpl} has been replaced by \code{\link{ppm}}.
}
\keyword{internal}
