\name{dim.detpointprocfamily}
\alias{dim.detpointprocfamily}
\title{Dimension of Determinantal Point Process Model}
\description{Extracts the dimension of a determinantal point process model.}
\usage{
  \method{dim}{detpointprocfamily}(x)
}
\arguments{
  \item{x}{object of class \code{"detpointprocfamily"}.}
}
\value{A numeric (or NULL if the dimension of the model is unspecified).}
\author{
  \adrian
  
  
  \rolf
  
  and \ege
  
}
\keyword{spatial}
\keyword{models}
