\name{is.dppm}
\alias{is.dppm}
\title{Recognise Fitted Determinantal Point Process Models}
\description{Check that an object inherits the class dppm}
\usage{is.dppm(x)}
\arguments{
  \item{x}{Any object.}
}

\value{A single logical value.}

\author{Ege Rubak <rubak@math.aau.dk>,
Adrian Baddeley <Adrian.Baddeley@uwa.edu.au>
and Rolf Turner <r.turner@auckland.ac.nz>}

\keyword{spatial}
\keyword{manip}
\keyword{models}
