\name{methods.ppx}
\Rdversion{1.1}
\alias{print.ppx}
\alias{plot.ppx}
\title{
  Methods for Multidimensional Space-Time Point Patterns
}
\description{
  Methods for printing and plotting a general multidimensional
  space-time point pattern.
}
\usage{
\method{print}{ppx}(x, ...)
\method{plot}{ppx}(x, ...)
}
\arguments{
  \item{x}{
    Multidimensional point pattern (object of class \code{"ppx"}).
  }
  \item{\dots}{
    Additional arguments passed to plot methods.
  }
}
\details{
  These are the \code{print} and \code{plot} methods for the
  class \code{"ppx"} of multidimensional point patterns.
}
\value{
  \code{NULL}.
}
\author{Adrian Baddeley
  \email{adrian@maths.uwa.edu.au}
  \url{http://www.maths.uwa.edu.au/~adrian/}
  and Rolf Turner
  \email{r.turner@auckland.ac.nz}
}
\seealso{
  \code{\link{ppx}}  
}
\keyword{spatial}
