\name{simplenet}
\alias{simplenet}
\docType{data}
\title{
  Simple Example of Linear Network
}
\description{
  A simple, artificially created, example of a linear network.
} 
\format{
  \code{simplenet} is an object of class \code{"linnet"}.
}
\usage{data(simplenet)}
\source{
  Created by Adrian Baddeley \email{Adrian.Baddeley@curtin.edu.au}.
}
\keyword{datasets}
\keyword{spatial}
