\name{spatstat-internal} 
\title{Internal spatstat functions}



\alias{print.autoexec}    
\alias{print.bugtable}

%%%%%%%
\description{
  Internal spatstat functions.
}
\usage{





\method{print}{autoexec}(x, \dots)
\method{print}{bugtable}(x, \dots)
}
\details{
  These internal \pkg{spatstat} functions should not be called
  directly by the user. Their names and capabilities may change
  without warning from one version of \pkg{spatstat} to the next.
}
\value{
The return values of these functions are not documented,
and may change without warning.
}
\keyword{internal}
