\name{update.detpointprocfamily}
\alias{update.detpointprocfamily}
\title{Set Parameter Values in a Determinantal Point Process Model}
\description{
  Set parameter values in a determinantal point process model object.
}
\usage{
  \method{update}{detpointprocfamily}(object, \dots)
}
\arguments{
  \item{object}{object of class \code{"detpointprocfamily"}.}
  \item{\dots}{
    arguments of the form \code{tag=value} specifying the parameters 
    values to set.
  } 
}
\author{
  \adrian
  
  
  \rolf
  
  and \ege
  
}
\keyword{spatial}
\keyword{models}




